adding 1.19.10 adjustments

+++ rke-cluster.sh
@@ -299,7 +299,8 @@ function _CONFIGURE_CUSTOM_CA {
         # server-ca.crt and key need to be copied only if not exist in target and if exist in source
         # only required on first master as others get it via registration against 9345
         # added first experiment for rancher and cert-manager using custom CA
-        if [ -f $CA_CRT ] && [ -f $CA_KEY ] && [ "$FIRSTMASTER" == "1" ]; then
+	# does not work with RKE2 v1.19.10+rke2r1
+        if [ -f $CA_CRT ] && [ -f $CA_KEY ] && [ "$FIRSTMASTER" == "1" ] && [ ! "$RKE2_VERSION" == "v1.19.10+rke2r1" ]; then
 	        sudo mkdir -p /var/lib/rancher/rke2/server/tls
                 echo 'custom CA exists and target CA does not exist - so copy custom one'
                 sudo cp -a $CA_CRT /var/lib/rancher/rke2/server/tls/server-ca.crt
@@ -330,6 +331,15 @@ function _CONFIGURE_CUSTOM_CA {
                         sudo mv $RKECLUSTERDIR/manifests/rancher.yaml.self-signed-ca $RKECLUSTERDIR/manifests/rancher.yaml
                 fi
         fi
+}
+
+function _FIX_1_19_10_DEPLOYMENT {
+	if [ "$RKE2_VERSION" == "v1.19.10+rke2r1" ]; then
+		# etcd timeout does not exist in 1.19.10
+		sudo sed -i "/^etcd-s3-timeout:.*/d" /etc/rancher/rke2/config.yaml
+		# cis 1.6 does not exist in 1.19.10
+		sudo sed -i "s/^profile: cis-1.6/profile: cis-1.5/g" /etc/rancher/rke2/config.yaml
+	fi
 }
 
 function _FIX_1_20_DEPLOYMENT {
@@ -391,7 +401,10 @@ function _FIX_1_20_6 {
 }
 
 function _FIX_1_20_7 {
-        if [ "$RKE2_VERSION" == "v1.20.7+rke2r1" ] || [ "$RKE2_VERSION" == "v1.20.7+rke2r2" ] || [ "$RKE2_VERSION" == "v1.20.8+rke2r1" ] || [ "$RKE2_VERSION" == "v1.21.2+rke2r1" ] ; then
+        if [ "$RKE2_VERSION" == "v1.20.7+rke2r1" ] ||\
+	   [ "$RKE2_VERSION" == "v1.20.7+rke2r2" ] ||\
+	   [ "$RKE2_VERSION" == "v1.20.8+rke2r1" ] ||\
+	   [ "$RKE2_VERSION" == "v1.21.2+rke2r1" ] ; then
                 echo "adding etcd user also on agents due to https://github.com/rancher/rke2/issues/1063"
                 sudo useradd -r -c "etcd user" -s /sbin/nologin -M etcd 2>&1 >/dev/null
         fi
@@ -554,6 +567,7 @@ function _JOIN_CLUSTER {
 		_COPY_MANIFESTS_AND_CHARTS
 		_ADJUST_CLUSTER_IDENTITY
 		_VSPHERE_CONFIG
+		_FIX_1_19_10_DEPLOYMENT
                 _FIX_1_20_DEPLOYMENT
                 _FIX_1_20_6
                 sudo sed -i "/^server/d" /etc/rancher/rke2/config.yaml
@@ -571,6 +585,7 @@ function _JOIN_CLUSTER {
 		_COPY_MANIFESTS_AND_CHARTS
 		_ADJUST_CLUSTER_IDENTITY
 		_VSPHERE_CONFIG
+		_FIX_1_19_10_DEPLOYMENT
                 _FIX_1_20_DEPLOYMENT
                 _FIX_1_20_6
                 sudo systemctl enable rke2-server.service 2>&1 >/dev/null;
@@ -586,6 +601,7 @@ function _JOIN_CLUSTER {
                 _PREPARE_RKE2_CLOUD_CONFIG
 		_ADJUST_CLUSTER_IDENTITY
 		_VSPHERE_CONFIG
+		_FIX_1_19_10_DEPLOYMENT
                 _FIX_1_20_DEPLOYMENT
                 _FIX_1_20_6
 		_FIX_1_20_7
